pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "self_evolving_agent"
version = "0.1.0"
authors = [
    { name="Gemini", email="gemini@google.com" },
]
description = "A self-evolving agentic framework"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mlflow",
    "ray[rllib]",
    "torch",
    "transformers",
    "sentence-transformers",
    "faiss-cpu",
    "hyperopt",
    "accelerate",
    "bitsandbytes",
    "safetensors",
    "datasets",
    "auto-gptq",
    "scikit-learn"
]

[project.urls]
"Homepage" = "https://github.com/pitt-musa/self-evolving-agent"
"Bug Tracker" = "https://github.com/pitt-musa/self-evolving-agent/issues"
